requirements.txt
amqp==2.6.1
bcrypt==3.1.7
billiard==3.6.3.0
bootstrap4==0.1.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.9
django-anymail==8.2
django-discover-runner==1.0
django-heroku==0.3.1
django-mail-templated==2.6.5
django-rest-swagger==2.2.0
django-role-permissions==3.1.1
django-smart-selects==1.5.9
djangorestframework==3.12.2
idna==2.10
importlib-metadata==2.1.1
itypes==1.2.0
Jinja2==2.11.3
kombu==4.6.11
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
openapi-codec==1.3.2
Pillow==7.2.0
protobuf==3.14.0
psycopg2==2.7.7
pycparser==2.20
PyMsgBox==1.0.9
PyMySQL==0.10.1
PyMysqlDB==0.0.2
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
reportlab==3.5.59
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
vine==1.3.0
whitenoise==5.2.0
zipp==1.2.0
gunicorn==20.0.4
